Revert "  feat: add comprehensive Drizzle ORM error handling utilities and update drizzle packages (#493)"
This reverts commit ca2c0188718ac1cc7ce231e27713e06892ceaa66.
