+++ apps/backend/package.json
@@ -36,7 +36,7 @@
     "@ts-rest/nest": "^3.52.1",
     "axios": "^1.9.0",
     "cookie-parser": "^1.4.7",
-    "drizzle-zod": "^0.8.2",
+    "drizzle-zod": "^0.5.1",
     "http-status-codes": "^2.3.0",
     "reflect-metadata": "^0.2.2",
     "rxjs": "^7.8.2",
