+++ packages/database/package.json
@@ -39,7 +39,7 @@
   "license": "ISC",
   "dependencies": {
     "dotenv": "^16.5.0",
-    "drizzle-orm": "^0.44.4",
+    "drizzle-orm": "^0.41.0",
     "postgres": "^3.4.5",
     "zod": "catalog:"
   },
@@ -47,7 +47,7 @@
     "@repo/eslint-config": "workspace:*",
     "@repo/typescript-config": "workspace:*",
     "@types/node": "catalog:",
-    "drizzle-kit": "^0.31.4",
+    "drizzle-kit": "^0.30.6",
     "eslint": "catalog:",
     "tsx": "^4.19.3",
     "typescript": "catalog:"
