+++ packages/database/src/index.ts
@@ -1,4 +1,3 @@
 export * from "drizzle-orm";
-export * from "drizzle-orm/errors";
 export * from "./database";
 export * from "./schema";
